+++ src/generics.rs
@@ -2,9 +2,9 @@ pub mod constants;
 
 mod alignment;
 mod length;
-mod num_primitive;
+mod primitive_traits;
 mod scalar;
 pub use alignment::*;
 pub use length::*;
-pub use num_primitive::*;
+pub use primitive_traits::*;
 pub use scalar::*;