+++ src/quaternion/float.rs
@@ -1,4 +1,6 @@
-use crate::{Alignment, EulerRot, Matrix, PrimitiveFloat, Quaternion, Vector};
+use crate::{
+    Alignment, EulerRot, Matrix, PrimitiveFloat, Quaternion, Vector, utils::PrimitiveFloatUtils,
+};
 
 impl<T, A: Alignment> Quaternion<T, A>
 where
@@ -10,7 +12,7 @@ where
     #[inline]
     #[must_use]
     pub fn from_rotation_x(angle: T) -> Self {
-        let (sin, cos) = (angle * T::as_from(0.5)).sin_cos();
+        let (sin, cos) = PrimitiveFloatUtils::sin_cos(angle * T::as_from(0.5));
         Self::from_xyzw(sin, T::ZERO, T::ZERO, cos)
     }
 
@@ -20,7 +22,7 @@ where
     #[inline]
     #[must_use]
     pub fn from_rotation_y(angle: T) -> Self {
-        let (sin, cos) = (angle * T::as_from(0.5)).sin_cos();
+        let (sin, cos) = PrimitiveFloatUtils::sin_cos(angle * T::as_from(0.5));
         Self::from_xyzw(T::ZERO, sin, T::ZERO, cos)
     }
 
@@ -30,7 +32,7 @@ where
     #[inline]
     #[must_use]
     pub fn from_rotation_z(angle: T) -> Self {
-        let (sin, cos) = (angle * T::as_from(0.5)).sin_cos();
+        let (sin, cos) = PrimitiveFloatUtils::sin_cos(angle * T::as_from(0.5));
         Self::from_xyzw(T::ZERO, T::ZERO, sin, cos)
     }
 
@@ -49,7 +51,7 @@ where
     pub fn from_axis_angle(axis: Vector<3, T, A>, angle: T) -> Self {
         debug_assert!(axis.is_normalized());
 
-        let (sin, cos) = (angle * T::as_from(0.5)).sin_cos();
+        let (sin, cos) = PrimitiveFloatUtils::sin_cos(angle * T::as_from(0.5));
         let xyz = axis * sin;
         Self::from_xyzw(xyz.x, xyz.y, xyz.z, cos)
     }
@@ -64,7 +66,7 @@ where
         if angle == T::ZERO {
             Self::IDENTITY
         } else {
-            let (sin, cos) = (angle * T::as_from(0.5)).sin_cos();
+            let (sin, cos) = PrimitiveFloatUtils::sin_cos(angle * T::as_from(0.5));
             let xyz = scaled_axis / angle * sin;
             Self::from_xyzw(xyz.x, xyz.y, xyz.z, cos)
         }
@@ -178,9 +180,9 @@ where
         let ti = angles.x * T::as_from(0.5);
         let tj = angles.y * T::as_from(0.5);
         let th = angles.z * T::as_from(0.5);
-        let (si, ci) = ti.sin_cos();
-        let (sj, cj) = tj.sin_cos();
-        let (sh, ch) = th.sin_cos();
+        let (si, ci) = PrimitiveFloatUtils::sin_cos(ti);
+        let (sj, cj) = PrimitiveFloatUtils::sin_cos(tj);
+        let (sh, ch) = PrimitiveFloatUtils::sin_cos(th);
         let cc = ci * ch;
         let cs = ci * sh;
         let sc = si * ch;
@@ -236,7 +238,7 @@ where
             if dif10 <= T::ZERO {
                 // x^2 >= y^2
                 let four_xsq = omm22 - dif10;
-                let inv4x = T::as_from(0.5) / four_xsq.sqrt();
+                let inv4x = T::as_from(0.5) / PrimitiveFloatUtils::sqrt(four_xsq);
 
                 Self::from_xyzw(
                     four_xsq * inv4x,
@@ -247,7 +249,7 @@ where
             } else {
                 // y^2 >= x^2
                 let four_ysq = omm22 + dif10;
-                let inv4y = T::as_from(0.5) / four_ysq.sqrt();
+                let inv4y = T::as_from(0.5) / PrimitiveFloatUtils::sqrt(four_ysq);
 
                 Self::from_xyzw(
                     (m01 + m10) * inv4y,
@@ -264,7 +266,7 @@ where
             if sum10 <= T::ZERO {
                 // z^2 >= w^2
                 let four_zsq = opm22 - sum10;
-                let inv4z = T::as_from(0.5) / four_zsq.sqrt();
+                let inv4z = T::as_from(0.5) / PrimitiveFloatUtils::sqrt(four_zsq);
 
                 Self::from_xyzw(
                     (m02 + m20) * inv4z,
@@ -275,7 +277,7 @@ where
             } else {
                 // w^2 >= z^2
                 let four_wsq = opm22 + sum10;
-                let inv4w = T::as_from(0.5) / four_wsq.sqrt();
+                let inv4w = T::as_from(0.5) / PrimitiveFloatUtils::sqrt(four_wsq);
 
                 Self::from_xyzw(
                     (m12 - m21) * inv4w,
@@ -381,7 +383,7 @@ where
 
         if length >= T::as_from(1e-8) {
             let axis = xyz / length;
-            let angle = length.atan2(self.w) * T::as_from(2.0);
+            let angle = PrimitiveFloatUtils::atan2(length, self.w) * T::as_from(2.0);
 
             (axis, angle)
         } else {
@@ -489,7 +491,7 @@ where
     pub fn angle_between(self, other: Self) -> T {
         debug_assert!(self.is_normalized() && other.is_normalized());
 
-        self.dot(other).abs().min(T::ONE).acos() * T::as_from(2.0)
+        PrimitiveFloatUtils::acos(self.dot(other).abs().min(T::ONE)) * T::as_from(2.0)
     }
 
     /// Computes the linear interpolation between `self` and `other` based on
@@ -556,7 +558,7 @@ where
             // If above threshold, perform linear interpolation to avoid divide by zero.
             (self * (T::ONE - t) + other * t).normalize()
         } else {
-            let theta = dot.acos();
+            let theta = PrimitiveFloatUtils::acos(dot);
 
             let x = T::ONE - t;
             let y = t;
@@ -744,11 +746,16 @@ where
 
 #[cfg(test)]
 mod tests {
+    #[cfg(feature = "num-primitive")]
+    use num_primitive::PrimitiveFloat;
+
+    #[cfg(not(feature = "num-primitive"))]
+    use crate::utils::PrimitiveFloatUtils;
     use crate::{
         EulerRot, Matrix, QuatA, Quaternion, Vector,
         utils::{
-            PrimitiveFloatFns, assert_debug_panic, assert_panic_test_eq, assert_test_eq, for_types,
-            random_iter, test_eq,
+            assert_debug_panic, assert_panic_test_eq, assert_test_eq, for_types, random_iter,
+            test_eq,
         },
     };